final move, condensed App into index
